--- .ScratchPad/PersistDataTest/src/services/appService.ts
@@ -1,21 +1,0 @@
-import { ConfigurationData } from "../types/configurationData";
-
-
-
-
-// export const setConfigurationData = () => {
-
-    
-//     upsertConfigurationData(result)
-// }
-
-// export const getSupportedGenerations = () => {
-//     // return 
-//     return getConfigData()
-//         .map( (d: ConfigurationData) => {
-//             return {
-//                 start_dex_no: d.starting_dex_no,
-//                 count: d.count
-//             }
-//         })
-// }
